+++ src/web/assets/cp/src/js/BaseElementIndexView.js
@@ -86,7 +86,10 @@ Craft.BaseElementIndexView = Garnish.Base.extend(
       // Enable inline element editing if this is an index page
       if (this.elementIndex.isAdministrative) {
         this._handleElementEditing = (ev) => {
-          if ($(ev.target).closest('a[href],button,[role=button]').length) {
+          if (
+            this.elementIndex.inlineEditing ||
+            $(ev.target).closest('a[href],button,[role=button]').length
+          ) {
             // Let the link/button do its thing
             return;
           }